@@ -1,9 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>net48;net7.0</TargetFrameworks>
-    <LangVersion>latest</LangVersion>
+	  <TargetFrameworks>net48;net8.0</TargetFrameworks>
+	  <OutputType>Exe</OutputType>
+	  <LangVersion>latest</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
