@@ -74,7 +74,7 @@ public struct DoubleStruct : IEquatable<DoubleStruct>
     /// <summary>Determines whether the specified <see cref="object"/>, is equal to this instance.</summary>
     /// <param name="obj">The <see cref="object"/> to compare with this instance.</param>
     /// <returns><c>true</c> if the specified <see cref="object"/> is equal to this instance; otherwise, <c>false</c>.</returns>
-    public override readonly bool Equals(object obj) => obj is DoubleStruct ds && Equals(ds);
+    public override readonly bool Equals(object? obj) => obj is DoubleStruct ds && Equals(ds);
 
     /// <summary>Determines whether the specified <see cref="DoubleStruct"/>, is equal to this instance.</summary>
     /// <param name="other">The <see cref="DoubleStruct"/> to compare with this instance.</param>
