@@ -2,30 +2,28 @@
 
 namespace Cave.IO;
 
-/// <summary>
-/// Attribute for marking fields or properties for serialization into ini file sections.
-/// </summary>
+/// <summary>Attribute for marking fields or properties for serialization into ini file sections.</summary>
 [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false)]
 public class IniSectionAttribute : Attribute
 {
-    /// <summary>
-    /// Initializes a new instance of the <see cref="IniSectionAttribute"/> class.
-    /// </summary>
+    #region Public Constructors
+
+    /// <summary>Initializes a new instance of the <see cref="IniSectionAttribute"/> class.</summary>
     public IniSectionAttribute() { }
 
-    /// <summary>
-    /// Initializes a new instance of the <see cref="IniSectionAttribute"/> class.
-    /// </summary>
+    /// <summary>Initializes a new instance of the <see cref="IniSectionAttribute"/> class.</summary>
     /// <param name="name"></param>
     public IniSectionAttribute(string name) => Name = name ?? throw new ArgumentNullException(nameof(name));
 
-    /// <summary>
-    /// Gets the section name.
-    /// </summary>
-    public string Name { get; set; }
+    #endregion Public Constructors
 
-    /// <summary>
-    /// Gets or sets the type of elements to be serialized.
-    /// </summary>
+    #region Public Properties
+
+    /// <summary>Gets the section name.</summary>
+    public string? Name { get; set; }
+
+    /// <summary>Gets or sets the type of elements to be serialized.</summary>
     public IniSettingsType SettingsType { get; set; }
+
+    #endregion Public Properties
 }
